Shifted pointerize_args_and_kwargs
