+++ src/syft/core/pointer/pointer.py
@@ -88,13 +88,12 @@ Example:
 import time
 
 # from typing import Callable as CallableT
-# from typing import Dict
 from typing import Any
+from typing import Dict
 from typing import List
 from typing import Optional
-
-# from typing import Tuple
-# from typing import Union
+from typing import Tuple
+from typing import Union
 import warnings
 
 # third party
@@ -599,6 +598,9 @@ class Pointer(AbstractPointer):
             self.client.gc.apply(self)
 
 
+# Taken from syft/core/ast/klass.py
+
+
 def _resolve_pointer_type(self: Pointer) -> Pointer:
     """
     Creates a request on a pointer to validate and regenerate the current pointer type. This method
@@ -639,3 +641,32 @@ def _resolve_pointer_type(self: Pointer) -> Pointer:
     del self
 
     return new_pointer
+
+
+def pointerize_args_and_kwargs(
+    args: Union[List[Any], Tuple[Any, ...]], kwargs: Dict[Any, Any], client: Any
+) -> Tuple[List[Any], Dict[Any, Any]]:
+    # When we try to send params to a remote function they need to be pointers so
+    # that they can be serialized and fetched from the remote store on arrival
+    # this ensures that any args which are passed in from the user side are first
+    # converted to pointers and sent then the pointer values are used for the
+    # method invocation
+    pointer_args = []
+    pointer_kwargs = {}
+    for arg in args:
+        # check if its already a pointer
+        if not isinstance(arg, Pointer):
+            arg_ptr = arg.send(client, pointable=False)
+            pointer_args.append(arg_ptr)
+        else:
+            pointer_args.append(arg)
+
+    for k, arg in kwargs.items():
+        # check if its already a pointer
+        if not isinstance(arg, Pointer):
+            arg_ptr = arg.send(client, pointable=False)
+            pointer_kwargs[k] = arg_ptr
+        else:
+            pointer_kwargs[k] = arg
+
+    return pointer_args, pointer_kwargs
